# api-gateway/src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Social Media Microservices API
  description: API documentation for the Social Media Microservices platform
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: /v1
    description: API Gateway
...
paths:
...
  /posts:
    get:
      tags:
        - Posts
      summary: Get posts feed
      description: Get a paginated list of posts
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of posts per page
          schema:
            type: integer
            default: 10
        - name: sort
          in: query
          description: Sort order
          schema:
            type: string
            enum: [recent, popular]
            default: recent
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostsListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    post:
      tags:
        - Posts
      summary: Create a new post
      description: Create a new post with content and optional media attachments
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  example: "This is a sample post content"
                mediaIds:
                  type: array
                  items:
                    type: string
                  example: ["media123", "media456"]
                visibility:
                  type: string
                  enum: [public, friends, private]
                  default: public
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Media Service Routes
...
  /search:
    get:
      tags:
        - Search
      summary: Search content
      description: Search posts, users, and other content
      parameters:
        - name: q
          in: query
          description: Search query
          required: true
          schema:
            type: string
        - name: type
          in: query
          description: Type of content to search
          schema:
            type: string
            enum: [all, posts, users]
            default: all
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Results per page
          schema:
            type: integer
            default: 10
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Invalid query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Health Check Routes
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PostResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            id:
              type: string
              example: 60d21b4667d0d8992e610c85
            content:
              type: string
              example: "This is a sample post content"
            author:
              type: object
              properties:
                id:
                  type: string
                  example: 60d21b4667d0d8992e610c85
                username:
                  type: string
                  example: johndoe
            media:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    example: media123
                  type:
                    type: string
                    example: image
                  url:
                    type: string
                    example: https://example.com/media/image.jpg
            visibility:
              type: string
              example: public
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    PostsListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            posts:
              type: array
              items:
                $ref: '#/components/schemas/PostResponse/properties/data'
            pagination:
              type: object
              properties:
                page:
                  type: integer
                  example: 1
                limit:
                  type: integer
                  example: 10
                totalDocs:
                  type: integer
                  example: 100
                totalPages:
                  type: integer
                  example: 10
                hasNextPage:
                  type: boolean
                  example: true
                hasPrevPage:
                  type: boolean
                  example: false
...
    SearchResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            results:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    example: 60d21b4667d0d8992e610c85
                  type:
                    type: string
                    example: post
                  content:
                    type: string
                    example: "This is a sample post content"
                  author:
                    type: object
                    properties:
                      id:
                        type: string
                        example: 60d21b4667d0d8992e610c85
                      username:
                        type: string
                        example: johndoe
                  createdAt:
                    type: string
                    format: date-time
            pagination:
              type: object
              properties:
                page:
                  type: integer
                  example: 1
                limit:
                  type: integer
                  example: 10
                totalResults:
                  type: integer
                  example: 100
                totalPages:
                  type: integer
                  example: 10
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Error message" 